Refactoring on connection to the logs database
The idea is to allow multiple databases to be connected to it.

The idea, for the logs, is allowing other databaases (like something
with user configuration) to influence on how the logs are read, but
without changing the logs database itself.

This is a long commit and result of many trials and fails and learnings
of how SQLite works (it's an amazing piece of software, btw).