Add basic support for browser automated tests #104
It uses Gauge+Taiko+Chrome for browser based acceptance tests.

There are no tests implemented yet, just one used to check that the
whole suite really works.

In the future we'll need to test on firefox as well, which is open
source and less likely to diverge from open web standards.