Port batch log cleaner to python, improving its interface
Somehow relates to #105

The user can point to a directory with postfix logs (compressed or not)
and the tool will generate a tarball with all the cleaned logs and
modification date of the original files intact.

By default the tool does not read the whole files, but just the first
and last five lines.

Optionally the user can pass a --complete/-c option that will read and
clean the complete files.

Usage: ./tools/batch_log_cleaner.py -i /var/log/ -o archive.tar.gz