Prevent deleting postfix ip address setting when language is set during login process
During login, the chosen language in the language selector is sent to
the settings, and when the remaining settings (for now postfix IP
address) is not set, the current value stored in the database should be
used instead.

A new dependency, mergo, is used to perform handling how old and new
settings are combined.