Do not crash when tracking field is missing on deliverydb
Log result instead, and continue.

NOTE: this is a workaround that needs to be fixed as soon as possible,
to provide more precise results and not drop messages.

WIP: Improve some error handling on tracking

WIP: fix on buggy missing keys on the delivery table

WIP: json print delivery try on error